requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contextlib2==21.6.0
cryptography==40.0.1
cssselect2==0.7.0
DateTime==5.1
deprecation==2.1.0
Django==4.2
exceptiongroup==1.1.1
execnet==1.9.0
fonttools==4.39.3
glob2==0.7
html5lib==1.1
idna==3.4
img2pdf==0.4.4
iniconfig==2.0.0
lxml==4.9.2
mock==5.0.2
numpy==1.24.2
opencv-python==4.7.0.72
oscrypto==1.3.0
packaging==23.1
pandas==2.0.1
path==16.6.0
path.py==12.5.0
pdf417decoder==1.0.8
pikepdf==7.2.0
Pillow==9.5.0
pluggy==1.0.0
pycparser==2.21
pydyf==0.6.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pylibdmtx==0.1.10
pypdf==3.7.1
pyphen==0.14.0
pypng==0.20220715.0
pytest==7.3.1
pytest-shutil==1.7.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzbar==0.1.9
qrcode==7.3.1
qrtools==0.0.2
reportlab==3.6.12
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
termcolor==2.3.0
tinycss2==1.2.1
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
weasyprint==58.1
webencodings==0.5.1
xhtml2pdf==0.2.9
XlsxWriter==3.1.0
zope.interface==6.0
zopfli==0.2.2
zxing==1.0
